refactored and added functionality
decided to use the Test document as the Result document - added isComplete and isResult flags to Test. Added answer shuffling to each question on the backend. Submitting answers now updates the context and updates the mongodb atlas database. An icon is now displayed when the user submits an answer to indicate if their answer is correct or not. User can refresh mid test and test will be refetcheed. When the last question of a test is submitted, the isComplete flag is set to true. User cannot navigate to the next question until the current question has been answered, but can navigate all questions for which answers have been submitted. If user navigates to randomized-practice route using browser buttons or by typing url directly, we now handlle the lack of test data gracefully.