added data model for user's question history
added a model to store user's interaction with each question. Records: if user got it correct the last time they answered it, number of times question seen, and if the user has flagged the question. Also, added a controller with a working createUserQuestionData controller, routes. Tested in Postman, and all workingn fine.