user's question history functionality added
If a user completes a test, the number of times the user has seen each question in the test is increased by 1. If a user answers a question correctly or incorrectly, the 'lastAnswerCorrect' field of the UserQuestionData is set to true or false respectively. If a user flags or unflags a question, this is reflected in their UserQuestionData and also updated in the Test document they are working through.