bug fix and keeping UserQuestionData up to date solved
When a user creates a new Test, the questions are randomly selected from the Questions collection. At test creation I have added a check to see if each test question is in UserQuestionData. If not, it is added. Added 11 more questions. Fixed issue with question category not being saved in the database when a question was added.