Extra features and fixes
Fixed issue where user submits the final answer of a test and the test is set to complete. Needed to create and send back the results to the frontend and set the global context. The category section of the Results page is complete. The user can view whether they passed or failed, how many questions they answered correctly out of the total number of questions, and they can view how many questions they got correct for each category of question. Added loading spinner and NoData to the results page. Added ResultsButtons and TestButtons components to make conditional rendering of different buttons easier. These components are rendered in the TestFooter depending on the location.pathname.