various minor updates
Changed some colors for icons to improve contrast. Updated modal alert so that it displays different messages or no message at all depending on circumstances. If no test data exists, hitting the X button in a test will simply navigate back to the practice page in the dashboard. If a test is incomplete, user gets warned that progress will be lost and can continue or cancel. If a test is complete, user is asked if they want to continue or not. ModalAlert is now behaving perfectly, but there is code duplication in ResultsSharedLayout and Test – still needs work. Removed all code related to ModalAlert from TestNav. Updated updateUser controller and the user can now save changes to their profile. Added ‘role’ property to user model – default is ‘user’. Options are user and admin. Added logic to use Ethereal Email for test env verification emails, SendGrid otherwise. Added a test user to allow app to be demoed without signing up.