Added functionality to create UserQuestionData on login
Created a userQuestionData property on User model. The UserQuestionData document will be created upon logging in. If user.userQuestionData is true, we do nothing, if false then we invoke our new function, createUserQuestionData, to create a UserQuestionData document for the user. As a reminder,   UserQuestionData stores the interaction history of the user with questions (number of times seen, did they answer the question correctly when last seen, have they flagged it).