Working on results overview functionality on back and frontend
Added OverviewListItem component that is similar to CategoryListItem. Made it more generalized. Used code from Categories in Overview page. Added getOverviewResults controller to the userQuestionDataController. Added an ‘overview’ get route to the userQuestionDataRoute, which uses getOverviewResults. Manually checked my ‘unseen’ questions calculation and it appears good. For the Testuser, they have 34 questions in their UserQuestionData document. The total number of questions in the app is 35, so they have 1 unseen question. In addition, they have 22 questions where numTimesAnswered is 0. The total unseen questions is therefore 23.