Toggling flagged status in old Tests
Updated toggleFlagged in testController to accept a test ID as parameter. Updated toggleIsFlagged in appContext to use a Test ID.  Updated TestFooter with useParams so we can pass the testId into toggleIsFlagged. Updated our getTest controller so that it will update the isFlagged status of questions in a Test Document when the user requests an old Test document by ID. When a user requests an old Test, the updatedAt value for the  Test and UserQuestionData are compared and if the update time/date on UserQuestionData is later by 10 seconds or more, the Test document has the isFlagged status updated for all of its questions. If the user flags or unflags a question while browsing an old Test, the Test is first updated and then the UserQuestionData is updated with the new flagged status. If the user then refreshes, by allowing a 10 second grace period, we can avoid getTest iterating over the Test again to update the flagged status.