Graphing and PreviousTests styling and functionality.
Updated CustomTooltip styles and content displayed. Rounded test scores to 1 decimal place. Centered the chart on the page and horizontal scrolling is working. Chart width is dynamic – the chart takes up 100% of the container if there are 8 or fewer data points, and then expands at 12.5% of the container width for each additional data point. Fixed TestListItems not displaying correctly (hidden by navbar) on the PreviousTests page.