Updated toggle flagged controller.
Updated disclaimer text slightly so it scans better. Updated the toggledFlagged controller questionController.js. This controller expects a questionId and a testId that may be null. If there is no testId, then we are updating a test that is in progress OR the user is reading the questions and has toggled the flagged status. Therefore, if a testId exists, we get the test and update the flagged status of the Test document. If testId is null, then we are updating a test in progress. We find a Test that has isResult set to false. If test references something at this point, then we know we are updating a Test in progress, we update the Test. Finally, we update the UserQuestionData and return the Test (even if it is null). When reading questions, we don’t need anything back from the server. When browsing old tests or working on a test in progress, we do need the Test to be returned. Tested and working fine. This is difficult to read, and perhaps the functionality should be separated out.