Fixed error fetching questions to read.
Fixed issue with reading questions. Our getQuestionsRead controller attempts to get the flagged status from UserQuestionData and set it on our questions that we pass to the frontend for the user to browse. I was attempting to access the isFlagged property on updateItem when updateItem could be null, ie it does not exist in UserQuestionData.