Practice by category functionality
Added state and event handlers so the user can select the number of questions to be tested on. Frontend functionality finished. User can start a test, answer questions, and view the results. Fixed: Link components on Read.js too narrow on mobile. Fixed: Question spacing on the Question List page when there are very few questions.