Updated showStats and createTest controllers.
showStats: Tests that have an isComplete value of false are now omitted. Note: Tests have always had the isComplete value set to true when the last question is answered. The stats for these Tests are only included in UserQuestionData at this point. createTest: The user can now create a Test based on the questions that they have seen the least.