Fix for bug while reading questions and loading fix.
Added useEffect to Read.js to reset the currentQuestion index in appContext to zero. Reading question 17 in one category, exiting and attempting to read a category with less than 17 questions was causing nothing to be rendered. Added SmallLoading component to be used where a small area of the screen is rendered with fetched data. Added SmallLoading to CategoryPractice.