Added to Timer functionality.
Added logic: In Question, which renders our Answer components, if isComplete is true, the onClick event for the Answer component does nothing – hence when the timer reaches zero, the user can no longer select or submit answers. Updated setTestComplete in appContext to make patch request to backend. Updated updateTest controller to handle requests to update the isComplete flag to true. The addition of the timer has introduced a number of errors when rendering, accessing or browsing results – need fixing!