Fixed various issues related to the introduction of the timer for mock/official style tests.
Fix: Issue with results not being available when timer runs out. Updated setTestComplete in appContext to make a patch request to the updateTest controller. The updated Test and results are now sent back to the frontend. Fixed: the user is allowed to/restricted from browsing forward through the questions as appropriate. This relies on completion status of test, whether the answer for the current question is null or not, and whether or not the user is already at the last question in the test.