Various bug fixes related to: grayed out 'next' button, creation of UserQuestionData, ghost extra questions thhat TestUser had.
Updated PreviousTestButtons component to correctly display grayed out ‘next’ button when browsing previous tests. Updated login controller so that we check for UserQuestionData every time the user logs in, if it doesn’t exist, we create. Previously we were looking for a flag of true/false on the User, but if we’ve deleted the UserQuestionData document directly from the database, this flag is meaningless. Had an issue where the TestUser had 6 extra questions in their UserQuestionData than existed in the system, this was causing ‘least seen’ type tests to return 6 less questions than the user specified. Deleted TestUser’s UserQuestionData and recreated it, and now it seems okay. Issue probably caused while writing and testing code to create tests, which involved identifying questions that had been randomly selected for the test, but were not in UserQuestionData (user had never seen) and adding them to UserQuestionData. Changed official style/mock test timer back to 45 minutes from 5 seconds.