CSS and bug fixes.
Removed extra text from landing page to improve mobile look. Updated createTest controller so that flagged questions are returned when the user attempts to create a test with flagged questions. Updated CategoryPractice – the ChooseNumQuestionsStartTest component is nowC only rendered when numOfQuestions is not null. Clicking in the button container does not now deselect the current selection. A border is now applied only to the button that is selected and not to the button container. Updated CSS for mobile only for CategoryPractice and ChooseNumQuestionsStartTest.