Updates to registration page for best practice.
Added ‘weak’ and ‘good’ text to password input based on whether or not the password is strong or not. Styled text based on that. Submit button on registration page is a lighter color when password is not strong to indicate that it is not usable. Removed various console logs. Added ‘eye’ icon to toggle password masking/unmasking.