Code splitting optimizations. React-toastify tweak. Minor CSS change.
Made all components and pages imported into App.js loadable – only loaded when required. Updated notifySuccess in appContext to allow toasts to be closed at dynamic times. notifySuccess takes a time in milliseconds after which the toast is dismissed. The “login success, redirecting...” toast now disappears before we redirect instead of lingering for a time after we have redirected to the dashboard.  Changed height of SmallSideBar to 90vh – the top and bottom appeared to be touching the edge of the viewport on mobile.